Time: 4 ms (98.56%) | Memory: 46.2 MB (93.94%) - LeetSync

+++ 451-sort-characters-by-frequency/sort-characters-by-frequency.java
@@ -19,7 +19,7 @@ class Solution {
             for(int i = 0 ; i < maxFreq ; i++){
                 res.append(maxChar);
             }
-            freq[maxChar] = 0; // marked it is visited;
+            freq[(int)maxChar] = 0; // marked it is visited;
         }
